Initial checkin.
The code in here does not work. It compiles and runs. But it's very slow and something is wrong in the training process. Still, I'm putting it here for posterity because there is a lesson in it.

Don't roll your own linear algebra code! I do think the matrix code I wrote produces the right numbers. But speed is a huge factor. You just can't compete with low level code taking advantage of specialized hardware. Or, to put it another way, graphics cards exist for a reason.